+++ .gitignore
@@ -117,14 +117,11 @@ celerybeat.pid
 *.sage.py
 
 # Environments
+.env
 .robo
-robo-react/
-.venv
 robo/
-venv/
 ROBO/
 robo.bak/
-venv.bak/
 pythonenv*
 
 # Spyder project settings
