+++ RobotixWeb/settings.py
@@ -6,6 +6,7 @@ Generated by 'django-admin startproject' using Django 3.1.7.
 from pathlib import Path
 import os
 import sys
+from decouple import config
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
 sys.path.insert(0, os.path.join(BASE_DIR, 'apps'))
@@ -14,10 +15,11 @@ sys.path.insert(0, os.path.join(BASE_DIR, 'apps'))
 # See https://docs.djangoproject.com/en/3.1/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = 'o!m3w+(x3q3_c9*bpe))v%5=8w_)&o_s4b*z=p-ij(#*^v%vob'
+SECRET_KEY = config('SECRET_KEY')
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
+
+DEBUG = config('DEBUG', default=False, cast=bool)
 
 ALLOWED_HOSTS = ['*',]
 
@@ -198,20 +200,19 @@ ACCOUNT_EMAIL_VERIFICATION = 'mandatory'
 
 #email
 EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
-EMAIL_HOST = 'smtp.gmail.com'
-EMAIL_USE_TLS = True
-EMAIL_PORT = 587
+
+EMAIL_HOST = config('EMAIL_HOST', default='localhost')
+EMAIL_USE_TLS = config('EMAIL_USE_TLS', default=False, cast=bool)
+EMAIL_PORT = config('EMAIL_PORT', default=25, cast=int)
 
 if DEBUG:
     EMAIL_HOST_USER = 'wandavision5432@gmail.com'
 else :
-    EMAIL_HOST_USER = 'robotixclub@nitrr.ac.in'
-    #EMAIL_HOST_USER = 'wandavision5432@gmail.com'
+    EMAIL_HOST_USER = config('EMAIL_HOST_USER', default='')
 
-EMAIL_HOST_PASSWORD = 'Nitrrobots16'
+EMAIL_HOST_PASSWORD = config('EMAIL_HOST_PASSWORD', default='')
 
-# EMAIL_HOST = 'localhost'
-# EMAIL_PORT = 1025
+#messages
 MESSAGE_STORAGE = 'django.contrib.messages.storage.session.SessionStorage'
 LOGOUT_REDIRECT_URL = '/'
