+++ requirements.txt
@@ -1,6 +1,7 @@
 amqp==2.6.1
 asgiref==3.3.1
 attrs==20.3.0
+autopep8==1.5.7
 bcrypt==3.2.0
 billiard==3.6.4.0
 blessed==1.18.0
@@ -52,12 +53,14 @@ pathspec==0.5.9
 Pillow==8.1.2
 psycopg2-binary==2.8.6
 pyasn1==0.4.8
+pycodestyle==2.7.0
 pycparser==2.20
 PyJWT==2.0.1
 PyNaCl==1.4.0
 pyrsistent==0.17.3
 python-crontab==2.5.1
 python-dateutil==2.8.1
+python-decouple==3.4
 python3-openid==3.2.0
 pytz==2021.1
 PyYAML==5.3.1
@@ -73,6 +76,7 @@ sqlparse==0.4.1
 tablib==3.0.0
 termcolor==1.1.0
 texttable==1.6.3
+toml==0.10.2
 uritemplate==3.0.1
 urllib3==1.25.11
 vine==1.3.0
